requirements.txt
einops
einops-exts
transformers==4.28
# torch==2.0.1
# torchvision
pillow
more-itertools
datasets
braceexpand
webdataset
wandb
nltk
scipy
inflection
sentencepiece==0.1.98
pycocoevalcap
pycocotools
open_clip_torch>=2.16.0


